vim-patch:9.0.0867: wildmenu redrawing code is spread out
Problem:    Wildmenu redrawing code is spread out.
Solution:   Refactor to move code together. (closes vim/vim#11528)

https://github.com/vim/vim/commit/d6e91385f0f7256aec8f70373c9e3399770d22e5
